refactor(projects): move formatProjectDate to lib/utils.ts
Per Ayush's review on PR #1220 round 3 (format.ts:1): check
app/lib/utils.ts before creating a feature-local utility; if generally
useful and missing, add it there. formatProjectDate (MMM d, yyyy with
contextual year drop) is general — not a projects-list specific concern
— so it belongs with the other shared date helpers.

- Added to frontend/packages/app/src/lib/utils.ts alongside the
  existing date-fns imports.
- Cells.tsx imports from `@/lib/utils` instead of the deleted
  `./format`.
- Deleted frontend/packages/app/src/pages/projects/list/format.ts
  (empty after the move).
